Move mdcat main to bin dir
Prepares for multiple CLI frontends.
